Unify app routers
