+++ src/RootApp.tsx
@@ -1,4 +1,3 @@
-import loadable from '@loadable/component';
 import { History } from '@remix-run/router';
 import { QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
@@ -8,21 +7,14 @@ import { ApiProvider } from 'hooks/useApi';
 import { WebConfigProvider } from 'hooks/useWebConfig';
 import { queryClient } from 'utils/query/queryClient';
 
-const StableAppRouter = loadable(() => import('./apps/stable/AppRouter'));
-const RootAppRouter = loadable(() => import('./RootAppRouter'));
+import RootAppRouter from './RootAppRouter';
 
 const RootApp = ({ history }: Readonly<{ history: History }>) => {
-    const layoutMode = localStorage.getItem('layout');
-    const isExperimentalLayout = layoutMode === 'experimental';
-
     return (
         <QueryClientProvider client={queryClient}>
             <ApiProvider>
                 <WebConfigProvider>
-                    {isExperimentalLayout ?
-                        <RootAppRouter history={history} /> :
-                        <StableAppRouter history={history} />
-                    }
+                    <RootAppRouter history={history} />
                 </WebConfigProvider>
             </ApiProvider>
             <ReactQueryDevtools initialIsOpen={false} />
