+++ src/apps/experimental/routes/routes.tsx
@@ -6,11 +6,13 @@ import ConnectionRequired from 'components/ConnectionRequired';
 import { toAsyncPageRoute } from 'components/router/AsyncRoute';
 import { toViewManagerPageRoute } from 'components/router/LegacyRoute';
 import { toRedirectRoute } from 'components/router/Redirect';
-import AppLayout from '../AppLayout';
 
 import { ASYNC_USER_ROUTES } from './asyncRoutes';
 import { LEGACY_PUBLIC_ROUTES, LEGACY_USER_ROUTES } from './legacyRoutes';
 import VideoPage from './video';
+import loadable from '@loadable/component';
+
+const AppLayout = loadable(() => import('../AppLayout'));
 
 export const EXPERIMENTAL_APP_ROUTES: RouteObject[] = [
     {
